services/web/requirements.txt
click==8.1.3
cryptography==40.0.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
maskpass==0.3.7
mock-alchemy==0.2.6
psycopg2-binary==2.9.4
pytest==7.3.1
pytest-flask-sqlalchemy==1.1.0
SQLAlchemy-serializer
